BoardController, FileController, SubjectController

+++ src/main/java/Refact/CalvinRefact/repository/dto/file/FileSimpleDto.java
@@ -7,9 +7,9 @@ public class FileSimpleDto {
     private Long id;
     private String original_name;
     private String save_name;
-    private int size;
+    private Long size;
 
-    public FileSimpleDto(Long id, String original_name, String save_name, int size) {
+    public FileSimpleDto(Long id, String original_name, String save_name, Long size) {
         this.id = id;
         this.original_name = original_name;
         this.save_name = save_name;
